+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [0.0.6] - 2021/06/20.
+* add ImageButton Widget and Sample
+* add Gradient Button Widget and Sample
+
 ## [0.0.5] - 2021/06/11.
 * add ListView LoadMore Widget and Sample
 * add Dialog Box Widget and update Cascader Sample with Dialog Box Widget